+++ backend/src/main/java/com/NBE3_4_2_Team4/global/security/AuthManager.java
@@ -1,6 +1,6 @@
 package com.NBE3_4_2_Team4.global.security;
 
-import com.NBE3_4_2_Team4.global.security.user.CustomUser;
+import com.NBE3_4_2_Team4.global.security.user.customUser.CustomUser;
 import com.NBE3_4_2_Team4.domain.member.member.entity.Member;
 import org.springframework.security.authentication.AuthenticationCredentialsNotFoundException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
