+++ backend/src/main/java/com/NBE3_4_2_Team4/global/security/oauth2/CustomOAuth2UserService.java
@@ -3,10 +3,10 @@ package com.NBE3_4_2_Team4.global.security.oauth2;
 import com.NBE3_4_2_Team4.domain.member.OAuth2RefreshToken.service.OAuth2RefreshTokenService;
 import com.NBE3_4_2_Team4.global.security.oauth2.userInfo.service.OAuth2UserInfoService;
 import com.NBE3_4_2_Team4.global.security.oauth2.userInfo.OAuth2UserInfo;
-import com.NBE3_4_2_Team4.global.security.user.CustomUser;
+import com.NBE3_4_2_Team4.global.security.user.customUser.CustomUser;
 import com.NBE3_4_2_Team4.domain.member.member.entity.Member;
 import com.NBE3_4_2_Team4.domain.member.member.service.MemberService;
-import com.NBE3_4_2_Team4.global.security.user.TempUserBeforeSignUp;
+import com.NBE3_4_2_Team4.global.security.user.tempUserBeforeSignUp.TempUserBeforeSignUp;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
